langchain_bible_app/requirements.txt
flask
langchain
langchain-openai
langchain-community
python-dotenv
openai